pyproject.toml
[tool.poetry]
name = "slash-command-bot-dpy-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
discord-ext-slash = "^0.5.0"
discord-py-slash-command = "^1.1.0"
python-chess = "^1.999"
CairoSVG = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
